Review feedback; moved profile translation code from pkg/profiles into profile_translator
